+++ src/components/Publisher.vue
@@ -40,8 +40,8 @@
             <q-checkbox style="display: flex;" color="dark" class="q-mt-sm q-mb-sm" v-model="config.options.dup" label="Duplicate flag"/>
             <q-collapsible v-if="version === 5" class="q-mt-sm q-mb-sm bg-grey-4" label="Properties">
               <q-checkbox color="dark" class="q-mt-sm q-mb-sm" v-model="config.options.properties.payloadFormatIndicator" label="Payload format indicator"/>
-              <q-input color="dark" type="number" v-model="config.options.properties.messageExpiryInterval" float-label="Message expiry interval" :error="!isNil(config.options.properties.messageExpiryInterval) && (config.options.properties.messageExpiryInterval < 0 || config.options.properties.messageExpiryInterval > 0xffffffff)"/>
-              <q-input color="dark" type="number" v-model="config.options.properties.topicAlias" float-label="Topic alias" :error="!isNil(config.options.properties.topicAlias) && (config.options.properties.topicAlias <= 0 || config.options.properties.topicAlias > 0xffff)"/>
+              <q-input color="dark" type="number" v-model="config.options.properties.messageExpiryInterval" float-label="Message expiry interval" clearable :clear-value="undefined" :error="!isNil(config.options.properties.messageExpiryInterval) && (config.options.properties.messageExpiryInterval < 0 || config.options.properties.messageExpiryInterval > 0xffffffff)"/>
+              <q-input color="dark" type="number" v-model="config.options.properties.topicAlias" float-label="Topic alias" clearable :clear-value="undefined" :error="!isNil(config.options.properties.topicAlias) && (config.options.properties.topicAlias <= 0 || config.options.properties.topicAlias > 0xffff)"/>
               <q-input color="dark" v-model="config.options.properties.responseTopic" float-label="Response topic"/>
               <q-input color="dark" type="textarea" :max-height="50" v-model="config.options.properties.correlationData" float-label="Correlation data"/>
               <div>